Fix permission attributes in GetDeploymentCredentials
